requirements.txt
dash
dash-ag-grid
Flask-SQLAlchemy
pydantic
sqlalchemy
pyyaml
